backend/requirements.txt
pyjwt
fastapi[standard]
bcrypt
uvicorn[standard]
mariadb
